Split OptionChain filters into OptionChain.Filters.cs
OptionChain is partial now: the class core keeps the constructors and Clone,
the universe filters and strategy pickers live in their own file.
